Update BarChart to enable horizontal and vertical orientation, removed RowChart.jsx
